% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacteria}
\alias{bacteria}
\title{Abundance of bacteria}
\format{
A data frame with 167 rows and 3 variables:
\describe{
  \item{value}{measured abundance}
  \item{colour}{preferred colour for display}
  \item{label}{taxon name}
}
}
\usage{
bacteria
}
\description{
Names and abundances of bacterial taxa as measured in a study of biofilms.
}
\keyword{datasets}
