% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/circleLayout.R
\name{circleLayout}
\alias{circleLayout}
\title{Find non-overlapping layout of circles.}
\usage{
circleLayout(xyr, xlim, ylim, maxiter = 1000, wrap = TRUE)
}
\arguments{
\item{xyr}{3-column matrix or data.frame (centre X, centre Y, radius)}

\item{xlim}{bounds in the X direction; either a vector for [xmin, xmax)
or a single value interpreted as [0, xmax)}

\item{ylim}{bounds in the Y direction; either a vector for [ymin, ymax)
or a single value interpreted as [0, ymax)}

\item{maxiter}{maximum number of iterations to attempt}

\item{wrap}{whether to allow coordinate wrapping across bounds. If `true`,
  coordinate wrapping results in a toroidal space; if `false`, ordinates
  are simply restricted to bounds.}
}
\value{
A list with components:
  \describe{
    \item{layout}{A 3-column matrix or data.frame (centre x, centre y, radius).}
    \item{niter}{Number of iterations performed.}
  }
}
\description{
Attempts to find a layout for a given set of circles
within a rectangle such that is no overlap between circles.
}

