% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleProgressiveLayout.R
\name{circleProgressiveLayout}
\alias{circleProgressiveLayout}
\title{Progressive layout algorithm}
\usage{
circleProgressiveLayout(x, sizecol = 1, sizetype = c("area", "radius"))
}
\arguments{
\item{x}{Either a vector of circle sizes, or a matrix or data frame
with one column for circle sizes.}

\item{sizecol}{The index or name of the column in \code{x} for circle sizes.
Ignored if \code{x} is a vector.}

\item{sizetype}{The type of size values: either \code{"area"} (default) 
or \code{"radius"}. May be abbreviated.}
}
\value{
A data frame with columns: x, y, radius. If any of the input size values
  were non-positive or missing, the corresponding rows of the output data frame
  will be filled with \code{NA}s.
}
\description{
Arranges a set of circles, which are denoted by their sizes, by consecutively 
placing each circle externally tangent to two previously placed circles while 
avoiding overlaps.
}
\details{
Based on an algorithm described in the paper:
\emph{Visualization of large hierarchical data by circle packing}
by Weixin Wang, Hui Wang, Guozhong Dai, and Hongan Wang. Published
in \emph{Proceedings of the SIGCHI Conference on Human Factors in Computing Systems},
2006, pp. 517-520 \url{https://doi.org/10.1145/1124772.1124851}

The implementation here was adapted from a version written in C by Peter Menzel:
\url{https://github.com/pmenzel/packCircles}.
}
\examples{
areas <- sample(c(4, 16, 64), 100, rep = TRUE, prob = c(60, 30, 10))
packing <- circleProgressiveLayout(areas)

\dontrun{

# Graph the result with ggplot
library(ggplot2)

dat.gg <- circleLayoutVertices(packing)

ggplot(data = dat.gg, aes(x, y, group = id)) +
  geom_polygon(colour = "black", fill = "grey90") +
  coord_equal() +
  theme_void()

}

}
