% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/circleVertices.R
\name{circleVertices}
\alias{circleVertices}
\title{Generate vertex coordinates for a circle.}
\usage{
circleVertices(xc, yc, r, npoints = 25)
}
\arguments{
\item{xc}{centre X}

\item{yc}{centre Y}

\item{r}{radius}

\item{npoints}{number of vertices}
}
\value{
A 2-column matrix of X and Y values.
}
\description{
Generates vertex coordinates for a circle with given centre and radius.
}
\seealso{
\code{\link{circlePlotData}}
}

