% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/circleLayout.R
\name{circleLayout}
\alias{circleLayout}
\title{Find non-overlapping layout of circles.}
\usage{
circleLayout(xyr, xbound, ybound, max.iter = 1000)
}
\arguments{
\item{xyr}{3-column matrix or data.frame (centre X, centre Y, radius)}

\item{xbound}{X upper limit (lower assumed 0)}

\item{ybound}{Y upper limit (lower assumed 0)}

\item{max.iter}{maximum number of iterations to attempt}
}
\value{
A list with components:
  \describe{
    \item{layout}{A 3-column matrix or data.frame (centre x, centre y, radius).}
    \item{niter}{Number of iterations performed.}
  }
}
\description{
Attempts to find a layout for a given set of circles
within a rectangle such that is no overlap between circles.
}

