% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/circlePlotData.R
\name{circlePlotData}
\alias{circlePlotData}
\title{Generate plotting data for circles.}
\usage{
circlePlotData(xyr, npoints = 25)
}
\arguments{
\item{xyr}{matrix or data.frame of circle data (x, y, radius)}

\item{npoints}{number of vertices to generate for each circle}
}
\value{
A data.frame with columns: id, x, y; where id is an
  integer identifier for each circle.
}
\description{
Takes a 3-column matrix or data.frame of circle data (centre x, centre y,
radius) and generates a data.frame of plotting data to use with ggplot.
}
\examples{
\dontrun{
## draw some random circles with ggplot
library(ggplot2)

xmax = 100
ymax = 200
rmax <- 20
N <- 100

## Random centre coordinates and radii
xyr <- data.frame(x=runif(N, 0, xmax),
                  y=runif(N, 0, ymax),
                  r=rbeta(N, 1, 5) * rmax)

## Get data for circle vertices
dat <- circlePlotData(xyr)

## Draw them
ggplot(dat, aes(x, y, group=id)) +
  geom_path(colour="steelblue") +
  coord_equal(xlim=c(0, xmax), ylim=c(0, ymax)) +
  theme_bw()
}
}
\seealso{
\code{\link{circleVertices}}
}

