% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleVertices.R
\name{circlePlotData}
\alias{circlePlotData}
\title{Generate a set of circle vertices suitable for plotting}
\usage{
circlePlotData(layout, npoints = 25, xyr.cols = 1:3, id.col = NULL)
}
\arguments{
\item{layout}{A matrix or data.frame of circle data (x, y, radius). May contain
other columns, including an optional ID column.}

\item{npoints}{The number of vertices to generate for each circle.}

\item{xyr.cols}{Indices or names of columns for x, y, radius (in that order).
Default is columns 1-3.}

\item{id.col}{Optional index or name of column for circle IDs in output. 
If not provided, the output circle IDs will be the row numbers of
the input circle data.}
}
\value{
A data.frame with columns: id, x, y; where id is the unique
  integer identifier for each circle.
}
\description{
This function is deprecated and will be removed in a future release. 
Please use \code{circleLayoutVertices} instead.
}
\seealso{
\code{\link{circleLayoutVertices}} \code{\link{circleVertices}}
}
