% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleRemoveOverlaps.R
\name{circleRemoveOverlaps}
\alias{circleRemoveOverlaps}
\title{Filters a set of circles to remove all overlaps}
\usage{
circleRemoveOverlaps(x, xysizecols = 1:3, sizetype = c("area", "radius"))
}
\arguments{
\item{x}{A matrix or data frame containing circle x-y centre coordinates
and sizes (area or radius).}

\item{xysizecols}{The integer indices or names of the columns in \code{x} for
the centre x-y coordinates and sizes of circles. Default is \code{c(1,2,3)}.}

\item{sizetype}{The type of size values: either \code{"area"} (default) or 
\code{"radius"}. May be abbreviated.}
}
\value{
A data frame with centre coordinates and radii of selected circles.
}
\description{
Filters a set of circles to remove all overlaps
}
