% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleVertices.R
\name{circleVertices}
\alias{circleVertices}
\title{Generate vertex coordinates for a circle}
\usage{
circleVertices(xc, yc, radius, npoints = 25)
}
\arguments{
\item{xc}{Value for centre X ordinate.}

\item{yc}{Value for centre Y ordinate.}

\item{radius}{Value for radius.}

\item{npoints}{The number of distinct vertices required.}
}
\value{
A 2-column matrix of X and Y values. The final row is a copy
  of the first row to create a closed polygon, so the matrix has
  \code{npoints + 1} rows.
}
\description{
Generates vertex coordinates for a circle given its centre coordinates
and radius.
}
\seealso{
\code{\link{circleLayoutVertices}}
}
